requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Computer vision
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Machine learning
torch>=2.0.0
segment-anything>=1.0
scikit-learn>=1.3.0
ultralytics>=8.0.0

# Hardware control
pyserial>=3.5
